backend/requirements.txt
fastapi
uvicorn
pillow
numpy
scikit-learn
python-multipart